Break components into es6 modules
