+++ src/components/plaintext.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const PlainText = React.createClass({
+  render: function() {
+    return <span>{this.props.text}</span>
+  }
+})
